# openapi.yaml
openapi: 3.0.3
info:
  title: SkillBridge API
  version: 1.0.0
  description: >
    API contract for SkillBridge based on the Week 3 endpoint list.
    This API provides expert discovery, consultation booking,
    logistics integration with Maji Website, and delivery tracking.
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
  /experts:
    get:
      tags: [Experts]
      summary: Return verified expert profiles
      operationId: getExperts
      parameters:
        - name: estate
          in: query
          schema:
            type: string
          example: Westlands
        - name: sort
          in: query
          schema:
            type: string
            enum: [rating, proximity]
          example: rating
        - name: lat
          in: query
          schema:
            type: number
            format: double
          example: -1.286
        - name: lng
          in: query
          schema:
            type: number
            format: double
          example: 36.817
      responses:
        "200":
          description: Experts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expert"
...
  /skills:
    get:
      tags: [Skills]
      summary: Return categorized skill offers
      operationId: getSkills
      parameters:
        - name: tag
          in: query
          schema:
            type: string
          example: Agronomy
      responses:
        "200":
          description: Skills retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
...
  /requests:
    post:
      tags: [Requests]
      summary: Create a consultation booking request
      operationId: createConsultationRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationRequest"
      responses:
        "201":
          description: Consultation request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Expert does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Expert not found
    Expert:
      type: object
      required:
        - id
        - name
        - estate
        - rating
        - latitude
        - longitude
      properties:
        id:
          type: integer
          example: 12
        name:
          type: string
          example: Grace Mwangi
        estate:
          type: string
          example: Westlands
        rating:
          type: number
          format: float
          example: 4.8
        latitude:
          type: number
          example: -1.286
        longitude:
          type: number
          example: 36.817
        bio:
          type: string
          example: Certified Agronomy mentor.
    Skill:
      type: object
      required:
        - id
        - name
        - category
      properties:
        id:
          type: integer
          example: 5
        name:
          type: string
          example: Soil Health
        category:
          type: string
          example: Agriculture
...
    ConsultationRequest:
      type: object
      required:
        - farmerName
        - topic
        - expertId
        - date
      properties:
        farmerName:
          type: string
          example: John Kamau
        topic:
          type: string
          example: Irrigation Systems
        expertId:
          type: integer
          example: 12
        date:
          type: string
          format: date-time
          example: "2026-09-20T10:00:00Z"
    RequestResponse:
      type: object
      required:
        - requestId
        - status
      properties:
        requestId:
          type: integer
          example: 501
        status:
          type: string
          example: Pending
...